# specs/001-enterprise-knowledge-rag/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Knowledge Intelligence RAG — Backend API
  version: 1.0.0
  summary: Evidence-grounded question answering over publicly accessible documentation.
  description: |
    An independent technical demonstration. Not affiliated with, sponsored by, or endorsed
    by Atlassian or any documentation publisher. Every source returned by this API links back
    to the original publisher page (FR-049, FR-050).

    ## Answering

    `POST /chat/stream` is the primary answering interface: it returns a Server-Sent Events
    stream of pipeline stage events. `POST /chat` returns the same result as a single JSON
    object. Both execute the identical pipeline; the stream is transport, not a different
    code path, so the evaluation runner uses the non-streaming form and measured metrics
    reflect the same work the user sees.

    ## Model reasoning

    No endpoint, response, or event in this API exposes model reasoning or internal
    deliberation (FR-034). Provider calls disable reasoning exposure explicitly; see
    research R-014. The response schemas below are exhaustive and closed
    (`additionalProperties: false`), so a new field cannot appear without a contract change.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /sources:
    get:
      tags: [sources]
      summary: List registered sources
      operationId: listSources
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: enabled
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Registered sources with live counts
          content:
            application/json:
              schema:
                type: object
                required: [items, total, limit, offset]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Source' }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
    post:
      tags: [sources]
      summary: Register a source and start its first crawl
      description: |
        The URL is validated against the address guard **before any connection is opened**
        (FR-044). A disallowed address is rejected, never fetched. Re-registering an existing
        `start_url` returns `200` with the existing source.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SourceCreate' }
      responses:
        '201':
          description: 'Registered; crawl started unless the request set `start_crawl: false`'
          content:
            application/json:
              schema:
                type: object
                required: [source]
                properties:
                  source: { $ref: '#/components/schemas/Source' }
                  crawl_job:
                    description: |
                      The job that will perform the crawl. **Null when the request set
                      `start_crawl: false`.**

                      *Amended 2026-10-03.* This field was required and non-null. That forced a
                      registration with `start_crawl: false` to invent a job, and the only
                      invented job that can exist — a `queued` row — holds the partial unique
                      index `crawl_jobs_one_active_per_source`, so the operator's next explicit
                      `POST /sources/{source_id}/crawl` was refused as "already running" against
                      a crawl that never began. A `null` here is the honest answer; a caller that
                      asked for no crawl gets no job to poll.
                    anyOf:
                      - { $ref: '#/components/schemas/CrawlJob' }
                      - { type: 'null' }
        '200':
          description: Already registered; the existing source is returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '400':
          description: The address was refused (`SSRF_BLOCKED` or `INVALID_URL`) or input failed validation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: The site's directives disallow the starting URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /sources/{source_id}:
    parameters:
      - $ref: '#/components/parameters/SourceIdPath'
    get:
      tags: [sources]
      summary: Retrieve a source
      operationId: getSource
      responses:
        '200':
          description: The source
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [sources]
      summary: Update a source, including enabling or disabling it
      operationId: updateSource
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SourceUpdate' }
      responses:
        '200':
          description: The updated source
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [sources]
      summary: Remove a source and all of its content
      description: |
        Removes the registry rows **and** the indexed vectors, so the content can no longer
        be retrieved (FR-031). Documents are tombstoned before their vectors are deleted, so
        an in-flight crawl cannot resurrect them (FR-056). Idempotent.
      responses:
        '204': { description: Removed }
        '404': { $ref: '#/components/responses/NotFound' }
  /sources/{source_id}/crawl:
    post:
      tags: [sources, crawl-jobs]
      summary: Request a re-crawl of a source
      operationId: crawlSource
      parameters:
        - $ref: '#/components/parameters/SourceIdPath'
      responses:
        '202':
          description: Crawl accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CrawlJob' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: A crawl for this source is already in progress; the message names the job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                alreadyRunning:
                  value:
                    error:
                      code: CRAWL_ALREADY_RUNNING
                      message: A crawl is already running for this source (job 8a93f0c2-1d4e-4a7b-9c3f-6b1e2d4a5f70). Wait for it to finish or cancel it.
                      request_id: 0f3c1d2a-6b4e-4c1f-9a77-2b8e5d0c4a11
                      details: { crawl_job_id: 8a93f0c2-1d4e-4a7b-9c3f-6b1e2d4a5f70 }
...
components:
  parameters:
...
    SourceIdPath:
      name: source_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - SSRF_BLOCKED
                - INVALID_URL
                - ROBOTS_DISALLOWED
                - UNAUTHORIZED
                - NOT_FOUND
                - CONFLICT
                - CRAWL_ALREADY_RUNNING
                - DOCUMENT_DELETED
                - RATE_LIMITED
                - PROVIDER_ERROR
                - PROVIDER_RATE_LIMITED
                - VECTOR_SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message: { type: string }
            request_id: { type: string, format: uuid }
            details:
              type: object
              additionalProperties: true
              description: Field names and enumerated values only. Never free-form user input.
...
    Source:
      type: object
      additionalProperties: false
      required: [id, name, start_url, allowed_domains, enabled, status, page_count, unit_count]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        start_url: { type: string, format: uri }
        allowed_domains: { type: array, items: { type: string } }
        product_domain: { type: [string, 'null'] }
        enabled: { type: boolean }
        status: { type: string, enum: [active, disabled, error] }
        page_count: { type: integer, description: Live documents only; deleted documents are excluded. }
        unit_count: { type: integer }
        last_crawl_at: { type: [string, 'null'], format: date-time }
        last_crawl_status: { $ref: '#/components/schemas/CrawlJobStatus' }
        created_at: { type: string, format: date-time }
    SourceCreate:
      type: object
      additionalProperties: false
      required: [name, start_url]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        start_url: { type: string, format: uri, maxLength: 2000 }
        product_domain: { type: [string, 'null'] }
        max_pages:
          type: integer
          minimum: 1
          maximum: 500
          default: 100
          description: Bounded crawl (FR-027).
        max_depth: { type: integer, minimum: 0, maximum: 5, default: 2 }
        delay_seconds:
          type: number
          minimum: 0.1
          maximum: 60
          default: 1.0
          description: Per-request delay between pages (FR-027, FR-028).
        start_crawl: { type: boolean, default: true }
    SourceUpdate:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        enabled: { type: boolean }
        product_domain: { type: [string, 'null'] }
        max_pages: { type: integer, minimum: 1, maximum: 500 }
        max_depth: { type: integer, minimum: 0, maximum: 5 }
        delay_seconds: { type: number, minimum: 0.1, maximum: 60 }
...
    CrawlJobStatus:
      type: string
      enum: [queued, running, completed, completed_with_errors, failed, cancelled]
    CrawlJob:
      type: object
      additionalProperties: false
      required: [id, source_id, target_url, status, requested_at, counters]
      properties:
        id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        target_url: { type: string, format: uri }
        status: { $ref: '#/components/schemas/CrawlJobStatus' }
        requested_at: { type: string, format: date-time }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        counters:
          type: object
          additionalProperties: false
          required: [discovered, processed, unchanged, skipped, failed]
          properties:
            discovered: { type: integer }
            processed: { type: integer }
            unchanged: { type: integer, description: Pages whose fingerprint matched; not re-indexed (SC-012). }
            skipped: { type: integer }
            failed: { type: integer }
        skipped_reasons: { type: object, additionalProperties: { type: integer } }
        error_code: { type: [string, 'null'] }
        error_message: { type: [string, 'null'] }
...